[REF] SEOS-1294: Use custom log format
With our current state, the we do not have functionality in place to interface with a RTC, so displaying a full date does not result in a useful log format. The custom format applied reduces the timestamp only to hh:mm:ss, logging the time that has passed since system start.